golf out CLI class
