+++ container.ts
@@ -4,7 +4,6 @@ import { Config } from './config';
 import { Host } from './host';
 import { Editor } from './editor';
 import { Executor } from './executor';
-import { Translator } from './translator';
 
 const tracer = getTracer('main');
 
@@ -13,7 +12,6 @@ export interface Container<H extends Host> {
   host: H;
   editor: Editor;
   executor: Executor;
-  translator: Translator;
 }
 
 export function container<H extends Host>(
@@ -23,14 +21,12 @@ export function container<H extends Host>(
   return tracer.spanSync('container', () => {
     const editor = new Editor();
     const executor = new Executor(config, editor, host);
-    const translator = new Translator(config, executor, host);
 
     return {
       config,
       host,
       editor,
       executor,
-      translator,
     };
   });
 }
